requirements.txt
Flask==2.0.1
Werkzeug==2.0.2
flask-cors==3.0.10
google-generativeai==0.3.1
python-dotenv==0.19.2
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
faiss-cpu
gunicorn==20.1.0
PyPDF2==3.0.1
numpy>=1.26.0
rank_bm25
